优化已知问题

+++ install/install.sh
@@ -12,7 +12,7 @@ linuxInstall(){
     fi
     # 检测并创建文件夹
     if [ ! -d "/usr/local/CySSHClient" ];then
-    mkdie -p /usr/local/CySSHClient
+      mkdir -p /usr/local/CySSHClient
     fi
 
     echo "cp -rf ${dqwz}/* /usr/local/CySSHClient/"
@@ -28,27 +28,35 @@ linuxInstall(){
     if [ -d "/usr/local/CySSHClient_v1" ];then
         echo "delete /usr/local/CySSHClient_v1"
         rm -rf /usr/local/CySSHClient_v1
+        rm -f /usr/bin/cyssh
     fi
 
     rm -f /usr/local/CySSHClient/install.sh
+
+    chown -R root:root /usr/local/CySSHClient/cyssh
+    chown -R root:root /usr/local/CySSHClient/cyscp
+    chown -R root:root /usr/local/CySSHClient/csc-server
 
     chmod +x /usr/local/CySSHClient/cyssh
     chmod +x /usr/local/CySSHClient/cyscp
     chmod +x /usr/local/CySSHClient/csc-server
 
-    ln -sf /usr/local/CySSHClient/cyssh /usr/bin/cyssh
-    ln -sf /usr/local/CySSHClient/cyscp /usr/bin/cyscp
+    ln -sf /usr/local/CySSHClient/cyssh /usr/local/bin/
+    ln -sf /usr/local/CySSHClient/cyscp /usr/local/bin/
 
     echo "初始化数据......"
     cd /usr/local/CySSHClient/
     ./csc-server -init
 
-    \cp -f /usr/local/CySSHClient/cyssh.service /etc/systemd/system/
-
-    systemctl enable cyssh
-    systemctl start cyssh
-    systemctl status cyssh --no-pager
-
+    if [ ! -f "/etc/systemd/system/cyssh.service" ];then
+      \cp -f /usr/local/CySSHClient/cyssh.service /etc/systemd/system/
+      systemctl enable cyssh
+      systemctl start cyssh
+      systemctl status cyssh --no-pager
+    else
+      systemctl restart cyssh
+      systemctl status cyssh --no-pager
+    fi
     echo "安装完毕!"
 }
